Change percent type (#44)

+++ src/main/java/com/rbkmoney/analytics/converter/DaoErrorDistributionsToResponseConverter.java
@@ -2,7 +2,6 @@ package com.rbkmoney.analytics.converter;
 
 import com.rbkmoney.damsel.analytics.ErrorDistributionsResponse;
 import com.rbkmoney.damsel.analytics.NamingDistribution;
-import com.rbkmoney.damsel.analytics.PaymentToolDistributionResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,8 +15,7 @@ public class DaoErrorDistributionsToResponseConverter {
         List<NamingDistribution> collect = namingDistributions.stream()
                 .map(paymentToolDistribution -> new NamingDistribution()
                         .setName(paymentToolDistribution.getName())
-                        //TODO after proto to double
-                        .setPercents(paymentToolDistribution.getPercent().longValue())
+                        .setPercents(paymentToolDistribution.getPercent())
                 ).collect(toList());
 
         return new ErrorDistributionsResponse()
